requirements.txt
berserk==0.10.0
certifi==2021.10.8
chardet==4.0.0
Deprecated==1.2.13
idna==3.3
lichs==1.3.1
ndjson==0.3.1
python-chess==1.999
requests==2.26.0
urllib3==1.26.5
wrapt==1.13.3
pytest==7.2.*
openai=0.27.2